setup configuration for using mongodb and apollo-graphql, in action: seed sample data to mongodb and graphql: query, mutation

+++ backend/src/graphql/typeDefs.ts
@@ -6,7 +6,7 @@ export const typeDefs = gql`
     title: String!
     image: String!
     address: String!
-    price: Float!
+    price: Int!
     numOfGuests: Int!
     numOfBeds: Int!
     numOfBaths: Int!
